Enable Google Java Format (#30)
Enables GJF and reformats all source accordingly. Leverages the GJF Gradle plugin. On CI, we verify that source files are formatted correctly. Locally, we provide a script config/hooks/pre-commit.template that, once installed as a pre-commit hook, fixes formatting at commit time.  We also modify the `gradlew` script to install the pre-commit hook if it is not already present.