Add check that `@NonNull` fields are not read before they are initialized (#58)
We check field reads in initializer blocks and constructors; we do not yet check any reads in initializer methods.

For a relevant read of field f, we check that f has been assigned on all control-flow paths leading to that read.  This requires checking previous initializer blocks, and also writes inside the same method and in "safe" initialization methods invoked before the read.