Add support for (a subset of) org.jetbrains.annotations.Contract (#129)
Currently, we can only reason about cases where the contract specifies that the return value of the method depends on the nullness value of a single argument. This means we can reason about rules like the following:

@Contract("null -> true")
@Contract("_, null, _ -> false")
@Contract("!null, _ -> false; null, _ -> true")
@Contract("!null -> !null")

In the last case, nullness will be propagated iff the nullness of the argument is already known at invocation.

We don't support cases where the antecedent of the clause involves multiple arguments, except in limited cases.

For example for @Contract("null, null -> true") we know nothing when the method returns true (because truth of the consequent doesn't imply truth of the antecedent), and if it returns false, we only know that at least one of the two arguments was non-null, but can't know for sure which one.