add UnannotatedClasses option (#160)
Previously, one could use the `UnannotatedSubPackages` to mark entire packages as unannotated, but there was no way to set a particular class within an annotated package to be treated as unannotated.  This is useful, e.g., to deal with a particular library class with a `@Nullable` return annotation that cannot yet be handled in the client code base.