Fix empty comment added when AutoFixSuppressionComment is not set. (#225)
When `-XepOpt:NullAway:SuggestSuppressions` is set, but `-XepOpt:NullAway:AutoFixSuppressionComment` is not set, the expected behavior is to suggest `@SuppressWarnings("NullAway")` where needed, but not add any comment to it.

Before this PR, however, we were adding an empty comment (e.g. `@SuppressWarnings("NullAway") /* */`). This fixes that and adds an unit test to check for a possible regression.