Fix AutoFix for MethodRefs returning null. (#229)
This fixes a case where `@SuppressWarnings(...)` annotations would 
be added on MethodRefs at the point where the MethodRef 
is introduced, rather than the enclosing method. This happened in cases
of invalid overriding of the return parameter compared to the functional
interface.

Also, fix typo: Add a space after `@SuppressWarnings(...)` fixes,
even if no AutoFixSuppressionComment is provided.