Pass nullability info on enclosing locals into dataflow analysis for lambdas and anonymous / local classes (#235)
Before, when a local from an enclosing scope was accessing within a capturing scope (lambda or anonymous / local class) NullAway would just assume the local was non-null, which could miss issues.  We fix the problem by maintaining an `EnclosingEnvironmentNullness` data structure that associates with each capturing scope nullability info on the enclosing locals.  This information is fed into the initial store when doing dataflow analysis for the capturing scope.