Enable excluded class annotations to (mostly) work on inner classes (#239)
Fixes #237 

We override `BugChecker.customSuppressionAnnotations`, which works for both top-level and inner classes.

The fix has one limitation, which is that it does not work for inner classes in the same module as the annotation declaration, due to the `customSuppressionAnnotations()` API requiring `Class` objects rather than matching by name.  We will just document this limitation as fixing is rather complex and potentially costly to performance.