Handle interaction between AcknowledgeRestrictiveAnnotations and TreatGeneratedAsUnannotated (#254)
If both the `TreatGeneratedAsUnannotated` and `AcknowledgeRestrictiveAnnotations` are set, previously we would still use any `@Nullable` annotations in generated code when checking clients.  This PR implements the desired behavior, which is to ignore those `@Nullable` annotations.

This points to a deeper issue in our terminology (kind of weird to use annotations from "unannotated" code) and config options.  But fully addressing that is outside the scope of this PR.