Disable JarInfer handler by default (#261)
This is mostly so that we can have a somewhat-functional NullAway on JDK 11 sooner (see #259).  But also, removing a handler may give a small performance boost.  With this change, you need to pass `-XepOpt:NullAway:JarInferEnabled=true` to enable JarInfer.

Also, we separate out the JarInfer tests into their own module, so that there is no direct dependence from the `nullaway` module to the `jarinfer` libraries.  This will enable us to get the `:nullaway:test` target passing on JDK 11 sooner.