Initial support for JDK 11 (#263)
For now, only the `:nullaway:test` target is passing on CI.  We need to address issues outlined in #259 to get everything working, and it may be a while before the sample Android app can easily be built.  That said, I think this initial level of support is useful and worth landing.