Revert "Model parseXXX methods of boxed primitives as having @NonNull parameter (#270)" (#281)
This reverts commit 33152f15f93dae9b219f902bdf18fb9ff0faeefc (#270).

See #269 for an explanation. Basically, `parseXXX` throws a `NumberFormatException` when passed `null`, and making it not able to take `null` means code that is correct now (because it catches `NumberFormatException`) would be broken in NullAway in a way that can't be resolved by simply adding nullness annotations or `castToNonNull`. This is, arguably, bad design on the part of all those `parseXXX` functions, but we have no control over that.