Allow models to override @Nullable on third-party functional interfaces (#326)
This change does two things, one uncontroversial and the other more opinionated

1) [Uncontroversial] It acknowledges `LibraryModels.nonNullParameters()` when 
deciding the nullability of the arguments to a functional interface which otherwise 
marks the parameters as `@Nullable`, in a few contexts where the functional 
interface is used to determine the nullability of the parameter inside a lambda or 
the feasibility of an unbound method reference usage. Before this change we 
did acknowledge library models when calling the FI methods directly, but not for 
e.g. lambda expression based overriding.

2) [Opinionated] We override the nullability of Guava's `Function` and `Predicate` 
functional interfaces. This means that lambdas implementing these in our code 
will need to return `@NonNull` but need not check the incoming arguments.
Ideally, we'd have generics support to specialize these interfaces for both lambdas 
that can take `null` and lambdas that don't, but this is the best we can likely do 
for now.

This has implications for checking Guava code with NullAway, but not for treating 
it as annotated vs unannotated with restrictive annotations enabled.