[JarInfer] Don't compute frames on bytecode writting. (#338)
Passing the `ClassWriter.COMPUTE_FRAMES` means that ASM must do things like 
resolve super types, which might involve classes neither in the processed jar, 
nor the classpath for the JarInfer tool. We don't actually need this flag as 
we are only adding annotations, not modifying the number or types or arguments 
or locals. If we ever do need to compute frame sizes, we might need to do so 
manually.