[JarInfer] Use exact jar output path when possible (#339)
With this, the `-o` parameter to JarInfer is always treated as the exact path
to an output jar file when the input (`-i`) is a single jar. When operating
on bytecode rewriting mode with multiple input jars, the path passed to `-o`
is treated as a directory. This allows the caller to know exactly where the
output jar will be written to on the first case, and still gracefully supports
the second.