Add option to JarInfer to deal with signed jars. (#345)
This makes two changes to the JarInfer CLI tool:

1) By default, the tool will error out when it encounters a signed jar 
(as per the `jarsigner` tool), since modifying a signed jar's bytecode 
and having it still verify is impossible. Users can then exclude such 
jars from the list of jars sent the tool runs on.
2) We add an optional flag `--strip-jar-signatures` for users who 
explicitly want to run JarInfer on such jars, removing the signature 
in the process.

This change includes a test, but it might be fragile outside of JDK8 
due to the fact that the `jarsigner` tool's programatic API is not 
stable. This instability does not affect the JarInfer tool, just the test.