Include outer$inner class name when reporting field init errors (#375)
Error message now includes the name of the class when reporting on
issues regarding `@NonNull` field initialization.

In particular, when dealing with inner classes, both the name of the
inner and outer class are included, e.g. `MyClass$InnerClass.f`, `A$1.f`,
`A$B$1.g`, etc. 

When multiple classes in the same file have fields with the same name, 
this  change makes it easier for developers to identify the right one, 
without having to match the line number in the error reported.

Tests included. Addresses #365