Improve matching of native Map methods (#390)
See Issue #389. 

NullAway detects and treats in a special way all methods from the `java.util.Map` interface. Before this change, the methods were identified only by simple name and by checking that the class (transitively) implements `Map`. However, this is a problem when implementing classes add new methods with the same name but different signature (e.g. number or types of arguments). 

This patch adds a check for the number of arguments. Subtyping and generics makes a check for the types of arguments involved and potentially expensive, so we don't do that for now, but will add it if the need arrises.

We include a test that would have triggered the crash from #389 in the previous NullAway commit.