Improve suppression of subcheckers, taking into account full AST path
This currently works for `@SuppressWarnings("NullAway.Optional"),
which will be acknowledged when placed:

* On a field being assigned a lambda which would otherwise trigger
  an error.
* On a method containing a line triggering an error.
* On a class which directly or indirectly contains the line triggering
  the error.

Included test cases show all the above scenarios.

We take use `state.getPath()` to make this determination, and it
involves a tree traversal, but since it should only happen where we
would otherwise output a NullAway error message, this should not be
high overhead.