Fix: android-jar.py's exit code. (#399)
This script would not propagate the exit code from
the final call to jarinfer, leading it to potentially
return success even when jarinfer fails to run.

This is not critical, since this script is only used for
the release process and other steps will fail before publishing
to maven central. Still worth fixing, though.

Credit to Clint Gibler, R2C, and the semgrep tool for finding
this.