Make excluded class annotations work on nested classes.
This fixes #401.

There is no way to check for this in the current EP
design without adding some overhead, since it's not
easy to know when we exit a nested class in order to
e.g. mantain a stack of nested classes and their
exclusion status (see comment in NullAway.java).

To minimize performance regressions, this takes the
following approach:

- Excluded class annotation in top-level classes are
  handled the same as before, setting up a compilation
  unit level flag and skipping all checking if that
  flag is true.
- Then, before generating an error, the AST path is
  traversed upwards, looking for any excluded class
  annotations. This is analogous to our handling of
  `@SuppressWarnings("NullAway.Optional")` and does
  add some overhead proportional to the error
  messages being suppressed by these annotations.

The advantage is that we expect no performance hit
on code that currently passes NullAway. The disadvantage
is that as user rely more on excluded class annotations
in nested classes, overhead can silently creep in.
I have no better solution for this without support
from core EP.

Note that this PR also generalizes the support we
already had for `@SuppressWarnings("NullAway.Init")`.
Now it is possible to put that annotation e.g. on
a method, and have it suppress initialization for
all nested classes created within the method.