Improved Map handling: Strings and integers.
This change improves our handling of map operations such as
`containsKey`, `put`, and `get`, by allowing keys that do
not have an access path directly associated with them.

In particular, it allows literal strings to be understood
as `Map` keys by NullAway for the purpose of matching
checked and accessed map-paths (fixing #398). It also
handles boxed integers and long integers, both as literals
and as primitive variables (fixing #410).