Try to handle lombok.Builder without crashing.
IMPORTANT:
This PR doesn't solve the fundamental incompatibilities between
lombok's code transformation approach and Error Prone's assumption
that the AST in memory corresponds to the original code for error
reporting.

What this change does is prevent NullAway for reporting on
@Builder generated code in most cases, while not completely
turning off checking at the top-level class scope.

We accomplish this via two separate hacks:
* First, we make @Builder a default external initialization
  annotation for the fields of the class in which it appears.
* Second, we add a handler plug-in point to suppress error
  messages before they are reported, and a LombokHandler
  to detect if we are inside the `FooBuilder` corresponding
  to a `@Builder Foo` class, and suppress all errors within.

The second step is needed, since `FooBuilder.f` will not be
covered by suppressing initialization warnings on `Foo`, and
since additionally Lombok neglects to mark `FooBuilder.f` as
`@Nullable` even if `Foo.f` is `@Nullable` (strangely, the
getter and constructor methods are created with the correct
nullability annotations, it seems).

We include tests and a new sample project running lombok
to more adequately test the interaction between NullAway and
the Lombok annotation processor for this and future changes.

We leave Lombok handling guarded by a flag, as it is likely
both slightly brittle and liable to have some performance
hit for code which uses Lombok (since it requires traversals
up the AST).