[Fix] Handle `WideningConversionNode` in Map key specifiers
This is a quick bug-fix to handle the case where the CFG node
being used as a Map key is a `WideningConversionNode`. This
node doesn't have it's own AST, instead using the `Tree` object
from its operand, which can cause Map key specifier detection
logic to crash if we don't check for it separatedly.