Try to handle lombok.Builder without crashing. (#414)
IMPORTANT:
This PR doesn't solve the fundamental incompatibilities between
lombok's code transformation approach and Error Prone's assumption
that the AST in memory corresponds to the original code for error
reporting.

What this change does is prevent NullAway for reporting on
`@Builder` generated code in most cases, while not completely
turning off checking at the top-level class scope.

We accomplish this by hard-coding the following two defaults
into NullAway:
* First, we make `@Builder` a default external initialization
  annotation for the fields of the class in which it appears.
* Second, we make `@lombok.Generated` an excluded
  class annotation, skipping analysis for classes marked
  as generated.

The use of the later class requires that lombok be configured
with `addLombokGeneratedAnnotation`. We add instructions
in the `README` to this effect.

We include tests and a new sample project running lombok
to more adequately test the interaction between NullAway and
the Lombok annotation processor for this and future changes.