GitHub Actions: Run CI on external PRs.
By default, GitHub Actions on push run only for the fork on which
the changes were pushed. Thus, if we want PRs from external contributors
to trigger CI, we need to trigger on `pull_request` in addition to
on `push`.

This might result on double jobs when the branch being used for the
PR is also in uber/NullAway, but so be it.

We also update coverage upload to happen only when the repo is
`uber/NullAway`. This avoids potential errors when the fork is
missing a `COVERALLS_REPO_TOKEN` secret.