GitHub Actions: Run CI on external PRs. (#442)
* GitHub Actions: Run CI on external PRs.

By default, GitHub Actions on push run only for the fork on which
the changes were pushed. Thus, if we want PRs from external contributors
to trigger CI, we need to trigger on `pull_request` in addition to
on `push`.

This might result on double jobs when the branch being used for the
PR is also in uber/NullAway, but so be it.

For the time being, we also allow coveralls to be flaky and don't block
the build on it, since we have been getting some issues when merging
PRs that don't have access to the `COVERALLS_REPO_TOKEN` secret,
plus general connection flakiness from the Coveralls gradle plugin. The
first issue could be solved by triggering on `pull_request_target`, but that
has security implications that can cause secrets to be leaked.