Fix publication of fat jar for jar-infer-cli.
The Gradle plugin switch from `maven` to `maven-publish` caused
by #457 broke our setup for publishing `jar-infer-cli` as a far jar.
Since the jarinfer CLI is used for processing arbitrary jars inside
other builds, including jars like guava that it depends on, it
must be a fat jar with no dependencies.

Unfortunatelly, so far, the best way we've found to make this work
is to skip the default `maven` publication, then make a new publication
using the result of the `shadow` plugin, and re-add to it the
sources and javadoc artifacts from the original `maven` publication.
We also had to manually set up the pom file settings for this custom
publication.