Fix publication of fat jar for jar-infer-cli. (#461)
The Gradle plugin switch from `maven` to `maven-publish` caused
by #457 broke our setup for publishing `jar-infer-cli` as a far jar.
Since the jarinfer CLI is used for processing arbitrary jars inside
other builds, including jars that it depends on (e.g. guava), it
must be a fat jar with no dependencies.

Unfortunately, so far, the best way we've found to make this work
is to skip the default `maven` publication, then make a new publication
using the result of the `shadow` plugin, and re-add to it the
sources and javadoc artifacts from the original `maven` publication.
We also had to manually set up the pom file settings for this custom
publication. It is bit of a hack, but the end result is a fat jar with a sane,
dependency-free, pom file, and passes SonaType repository closing 
validation.