Allow specifying custom names for Contract annotations (#476)
This adds a CLI parameter (`-XepOpt:NullAway:CustomContractAnnotations=`) to allow specifying names for Contract annotations to use instead of Jetbrains.

The structure of the spec string inside the annotation must still be the sub-set of Jetbrains' Contract spec currently supported by NullAway, and it must be available as the `value`/default argument to the annotation. The rationale is to support `@Contract` on codebases that do not wish to add the external Jetbrains dependency.

This implements #474.