Add a jmh module for benchmarking (#487)
This is a first crack at creating a jmh module for benchmarking NullAway, so we can more easily detect performance regressions.  We have a single benchmark for now, Caffeine 3.0.2.  `./gradlew jmh` runs the benchmark.  This is an MVP: in the future, we should add more benchmarks, make the benchmark runner more versatile, etc.

For now, running benchmarks only works on JDK 11.  We can add support for JDK 8 in the future if we want (by adding EP javac to bootclasspaths, etc.).