Run some tests on JDK 17 (#511)
This change adds a CI job so we run the same tests on JDK 17 that we currently run on JDK 11.  Currently we only test using the latest version of Error Prone; we could also test using 2.4.0 if desired.  In either case, the new job(s) should be added as "required" for future changes to land.