Test that DummyOptionsConfig's methods throw IllegalStateException
By design, this class is intended to fail whenever called, as an instance will only be generated when failing to produce an instance of `ErrorProneCLIFlagsConfig`.

See the javadoc documentation on `DummyOptionsConfig` itself as to why the class is needed to build NullAway.

Since we want NullAway to fail fast if this dummy options object is ever passed, we test that all declared methods in the class throw an exception if ever called. Our test uses reflection to account for future methods being added to the `Config` interface.