Add JarInfer models for Android SDK 31 (Android 12)
Generated based on AOSP tag android-12.0.0_r21.

This required minor changes to the RELEASING.md process and our
scripts, to account for changes in AOSP's build output structure.

Additionally, this PR includes a minor fix for the JarInfer
driver, which moves return-value inference code to the same
try-catch block for Wala IR generation errors as the parameter
nullability inference code.