Add task to run NullAway on itself (#542)
The task can be run as `./gradlew :nullaway:buildWithNullAway`.  This turned out to be less challenging than I expected.  Unfortunately, the NullAway implementation does not yet pass NullAway; we will need to address that before running this task on CI.  The task only gets created when building on JDK 11+; for reasons I don't understand, the task fails to configure on JDK 8.