Better fix for crash on member selects inside module-info.java (#544)
Now, in `matchMemberSelectTree()`, we check if the `MemberSelectTree` represents a reference to a module.  If it does, we do not check if its base expression could be `null`.  This change gives us a stronger invariant in `mayBeNullExpr()`: any `IdentifierTree` or `MemberSelectTree` passed to `mayBeNullExpr()` should have a non-null symbol.  We explicitly check this to help us flag any future weird cases more easily.

Unfortunately, to maintain JDK 8 compatibility, the check for whether a `MemberSelectTree` represents a module requires reflection.  Testing did not reveal any significant performance impact.  If / when NullAway requires JDK 11, we can and should get rid of the reflection.

Code coverage is reduced due to the new runtime checks for (hopefully) impossible cases in `mayBeNullExpr()`.

Fixes #533