Improve nullness tracking of map calls in the presence of type casts (#537)
When chains of `Map.get(a).get(b).get(c)` are used without casts, our existing map handling code suffices to keep track of non-null APs. However, this breaks in the presence of casts. Unfortunately, chains of maps with casts are a common way to handle hierarchical but untyped data sources, such as loaded JSON/Yaml.

In general, casts make tracking of access paths equivalence harder. As such, with this change, we strip all casts when considering the receiver of a method call or field access for access path construction. This seems to handle the   chained `Map<Object,Object>` pattern above, without being in any way specific to maps.