Allow zero-argument static method calls to be the root of an access path (#596)
This allows code like `if (System.console() != null) { System.console().readLine() }` to typecheck without storing `System.console()` in a local variable.  The treatment is unsound in principle, but we believe will rarely introduce unsoundness in practice.  (We have a similar unsoundness for zero-argument (ignoring the receiver) instance methods, which we assume are getters.)