Add config option to register custom `@Generated` annotations. (#600)
This is most useful in combination with the option to treat generated
code as unannotated. It's not enough to use `ExcludedClassAnnotations`,
unfortunately, since that will still assume the code is properly annotated.
It will simply forgo checking the implementation.

Also, remove redundant config methods to check for the value
of `-XepOpt:NullAway:TreatGeneratedAsUnannotated`.