Add NullAway 0.9.7 as a JMH benchmark (#602)
We now include NullAway 0.9.7 as a JMH benchmark for NullAway.  As part of this change, we significantly refactor the JMH code, to enable writing regression tests checking that the benchmarks compile.  (This was motivated by the fact that NullAway was a bit tricky to get compiling under JMH as it requires AutoValue.)  The logic for compiling the benchmarks has moved into the `*Compiler` classes, and the JMH harnesses are now just simple calls into those classes.