Cleanups to AccessPath representation and implementation (#603)
This cleans up how we represent the root of an access path.  Previously we used a `Root` class which contained an `Element`, with a boolean field indicating if the root was actually the receiver (`this`, which has no `Element`).  Now, we just make the root a `@Nullable Element`, with `null` representing the receiver.  Since we check our code with NullAway now this should be safe 🙂 And it's a slightly more efficient representation.

Alongside that change, clean up some of the messy logic inside `AccessPath` for access path building.  In doing this cleanup, I noticed that our streams handling didn't correctly handle cases involving a stream of `Map`s.  Not super realistic, but in any case it is fixed now and a test is added.