Treat Void formal arguments as @Nullable
This is a follow up on #595. See that PR for some of the soundness
caveats related to method overriding and the nullability of Void
in our current implementation.

This does, however, avoid some otherwise awkward suppressions,
as shown in #608.

There are two correct ways of hadling Void:

a) Having it default to `@Nullable` as a type, which requires us
   adding a "default nullability" for some types/classes and explictly
   contradicts JSpecify (jspecify/jspecify#51)
b) Supporting generics, and requiring explictly annotating `@Nullable Void`.
   This will probably also require third-party libraries which we consider
   annotated (such as CF), to adopt this convention.

I believe (b) is the way forward long-term, which means that, for now,
this hack might be the best we can do without generics support. Once
NullAway supports generics, both this change and #595 should be reverted.