Allow Library Models to override annotations. (#624)
Currently `LibraryModels` can only give information regarding **unannotated** source code. This PR enables `LibraryModels` to override annotations in an **annotated** source code. This behavior is controlled by a flag.

To enable this feature, the following error prone flag must be passed:

```shell
-XepOpt:NullAway:AllowLibraryModelsOverrideAnnotations=true
```

The main motivation for this PR is to prepare the foundation for [Annotator](https://github.com/nimakarimipour/NullAwayAnnotator) to perform upstream API analysis and is an alternative way for #620.