Allow tracking field accesses outside the this instance and static fields. (#625)
We took the decision, back around the origins of NullAway, to not track arbitrary field-dereferences for nullness propagation. Instead, NullAway expects field accesses to be: a) specifically on fields of the current instance object (i.e. `this`), b) on static fields (e.g. constants/enums), or c) made through getters.

This means that, for example, the following code would produce a surprising NullAway error:

```
if (foo.o == null) {
    foo.o = new Object();
}
return foo.o.toString();
```

Assuming field `o` of `Foo` is `@Nullable` and an instance field, the code above will cause NullAway to complain that `foo.o.toString()` might be dereferencing a "nullable" expression `foo.o`, despite the check right before it. This is because NullAway doesn't track any access paths of the form (argument/local)+field, whereas it does track (argument/local)+method call. This is unintuitive.

While getters are good practice in Java, they aren't universal. This is already a problem for JDK11 codebases and only likely to become more obvious when dealing with JDK17 records.

This PR attempts to relax this requirement.