Support for Map.computeIfAbsent(...) (#640)
This method should be treated as `Map.put(...)`, with the caveat that we need 
to check if the `Function<K, V>` passed to `computeIfAbsent(...)` can return a 
`@Nullable V`, rather than just checking for the nullness of the argument 
itself.

Without full support for nullability of type parameters, we can't check that we 
are dealing with `Function<K, V>` vs `Function<K, @Nullable V>` in the general 
case. For usability, we default to marking the map key access path as non-null 
if `computeIfAbsent(...)` is called for that key. We add a few sanity checks, 
but as the test cases show, those checks _can_ be incomplete. In practice, we 
expect `computeIfAbsent(...)` to compute a non-null value in all but extremely 
rare instances, but long-term it would be ideal if we could properly check this.