Generalize handler APIs for argument nullability on (un-)annotated code. (#639)
In particular, this allows library models to change the nullability of method
arguments in annotated code.

This change replaces two separate an overly specific `Handler` callbacks:

- `onUnannotatedInvocationGetExplicitlyNullablePositions`
- `onUnannotatedInvocationGetNonNullPositions`

With a single more general callback:

- `onOverrideMethodInvocationParametersNullability`

which is explicitly meant to handle overriding/changing method parameter 
nullability for both annotated and unannotated code (thus solving issue #637).

This requires some refactoring (mostly simplifying logic) in the core NullAway 
class, as well as changes to the different handlers, which must now explicitly 
skip annotated code if they aren't designed to work with it (e.g. 
`RestrictiveAnnotationHandler`).