Augment error serializarion info (#643)
This PR augments error information serialized by NullAway. At current state, if `SerializeFixMetada` is active, all reporting errors will be serialized to `errors.tsv` output file where each row has the format below:
```
MESSAGE_TYPE    MESSAGE   CLASS   METHOD
```
The above information does not contain any information regarding the `element` which involved in a pseudo-assignment of a `@Nullable` expression into a `@NonNull`. This information is crucial to detect origins of the error and other computable features.

With this PR, the format of the `errors.tsv` is changed to below to serialize all the required information to denote the involved element in the error:
```
message_type    message    enc_class    enc_method    kind    class   method     param    index    uri    
```



This PR involves other small refactoring (mostly renaming) as it was necessary to reuse the existing methods to produce the desired outputs.