Generalize handler APIs for return nullability on (un-)annotated code (#641)
This is a follow up to changes in #639, #636, and #624. Here we
generalize the former Handler callback:

- onUnannotatedInvocationGetExplicitlyNonNullReturn

which was used to change the nullability of method returns in
unannotated code for the purpose of method `@Override` checks
(i.e. subtyping).
    
We provide instead:
    
- onOverrideMethodInvocationReturnNullability
    
Which allows handlers to change the nullability of method
returns for both annotated and unannotated code, for the same
purposes.
    
Additionally, we implement this new handler callback in
`LibraryModelsHandler`.
    
Before this change, marking a return as non-null in a library model
affected checking of method invocations, but not necessarily
overriding/subtyping checks.