Optimizations for parameter nullness handler / overriding (#646)
Mostly (?) addresses #645.

Since this code seems to be performance critical, we should consider shifting our representation of parameter nullability from a `Map<Integer,Nullness>` to an `ArrayList<Nullness>`.  But that is a more intrusive change.  These changes seems be enough to address the performance regression from #639 (to be confirmed with a full run of perf regression tests).