Switch parameter overriding handler to use `Nullness[]` (#648)
The performance improvement from this change is very small at best, and it doesn't seem to fully remove the recent performance regression.  But, this change clearly makes the code more efficient in the common case, and I think a bit easier to read as well.  So we should keep it.  Any remaining performance regression (#645) is probably just due to doing more checking than we were doing before (to fix bugs), so we can tolerate it.

Fixes #645