[JSpecify] Support @NullMarked on methods. (#644)
This adds support for adding `@NullMarked` to methods, in addition to classes, 
which will be supported by the version of `@NullMarked` shipping with JSpecify 0.3.0.

Some relevant changes needed to accomplish this:

- We rename `ClassAnnotationInfo` to `CodeAnnotationInfo`. It still operates as a cache 
over classes and their annotations, but now each cache entry has a map with information 
about the null-markedness of its methods.
- We add a `checkForMethodNullMarkedness(...)` call to the beginning of `matchMethod(...)`, 
before we check for `withinAnnotatedCode(...)`. If we detect a `@NullMarked` annotation 
here, while otherwise looking at unmarked code, we switch our scan of the current top-level 
class to our slower `PARTIALLY_MARKED` scan mode.
- We add comprehensive tests to  `NullAwayJSpecifyTests.java` and make fixes in various 
places accordingly. An important change to note is that we need to sometimes compute 
`EnclosingEnvironmentNullness` information for anonymous or local classes that are 
otherwise unmarked but have a `@NullMarked` method, and we must do this inside
`checkForMethodNullMarkedness(...)` (since we don't know about the marked methods 
while processing `matchClass(...)`)